Create version specific upgrade mechanism.
